+++ packages/sdk/ui-react/src/widget/autoComplete/autoComplete.scss.d.ts
@@ -1,0 +1,5 @@
+// This is a generated file. Changes are likely to result in being overwritten
+export const autoCompleteContainer: string;
+export const selected: string;
+export const errorMessage: string;
+export const invalid: string;